+++ lib/ff_annotation_route.dart
@@ -1,9 +1,5 @@
 library ff_annotation_route;
 
-import 'dart:io' as io;
-
-import 'package:ff_annotation_route/src/utils/format.dart';
-
 import 'src/package_graph.dart';
 import 'src/route_generator.dart';
 
@@ -19,6 +15,7 @@ void generate(
   bool isPackage = false,
   bool routeSettingsNoIsInitialRoute = false,
   String outputPath,
+  String routesFileOutputPath,
 }) {
   RouteGenerator root;
   final List<RouteGenerator> nodes = <RouteGenerator>[];
@@ -33,26 +30,31 @@ void generate(
       routeGenerator.getLib();
       routeGenerator.scanLib();
       if (routeGenerator.hasAnnotationRoute) {
-        final io.File file = routeGenerator.generateFile();
-        formatFile(file);
+        //final io.File file =
+        routeGenerator.generateFile();
+        //formatFile(file);
         nodes.add(routeGenerator);
       }
     }
   }
   nodes.sort(
-    (RouteGenerator a, RouteGenerator b) => a.packageNode.name.compareTo(b.packageNode.name),
+    (RouteGenerator a, RouteGenerator b) =>
+        a.packageNode.name.compareTo(b.packageNode.name),
   );
   root?.getLib();
   if (rootAnnotationRouteEnable) {
     root?.scanLib(outputPath);
   }
-  final io.File routeFile = root?.generateFile(
+  //final io.File routeFile =
+  root?.generateFile(
     nodes: nodes,
     generateRouteNames: generateRouteNames,
     outputPath: outputPath,
     generateRouteConstants: generateRouteConstants,
+    routesFileOutputPath: routesFileOutputPath,
   );
-  final io.File helperFile = root?.generateHelperFile(
+  //final io.File helperFile =
+  root?.generateHelperFile(
     nodes: nodes,
     routeSettingsNoArguments: routeSettingsNoArguments,
     generateRouteHelper: generateRouteHelper,
@@ -60,8 +62,16 @@ void generate(
     outputPath: outputPath,
   );
 
-  formatFile(routeFile);
-  if (routeFile != null) {
-    formatFile(helperFile);
-  }
+   root?.generateHelperFile(
+    nodes: nodes,
+    routeSettingsNoArguments: routeSettingsNoArguments,
+    generateRouteHelper: generateRouteHelper,
+    routeSettingsNoIsInitialRoute: routeSettingsNoIsInitialRoute,
+    outputPath: outputPath,
+  );
+
+  // formatFile(routeFile);
+  // if (routeFile != null) {
+  //   formatFile(helperFile);
+  // }
 }