+++ lib/src/command/output.dart
@@ -3,7 +3,7 @@ import 'command.dart';
 class Output extends Command with CommandValue {
   @override
   String get description =>
-      'The main project route file is relative to the lib directory.';
+      'The path of main project route file and helper file.It is relative to the lib directory.';
 
   @override
   String get full => '--output';