+++ lib/src/command/route_constants.dart
@@ -1,6 +1,6 @@
 import 'package:ff_annotation_route/src/command/command.dart';
 
-class RouteConstants extends Command {
+class RouteConstants extends Command{
   @override
   String get description => 'Whether generate route names as constants.';