+++ lib/src/command/route_names.dart
@@ -11,4 +11,5 @@ class RouteNames extends Command {
   @override
   String get short => '-rn';
 
+
 }