+++ lib/src/utils/format.dart
@@ -1,5 +1,8 @@
 import 'dart:io';
+import 'package:dart_style/dart_style.dart';
 import 'package:io/ansi.dart';
+
+final DartFormatter formatter = DartFormatter();
 
 Future<void> formatFile(File file) async {
   if (file == null) {